+++ packages/web-components/src/components/content-group/__stories__/content-group.stories.ts
@@ -129,5 +129,20 @@ export default {
         cta: boolean('CTA:', true),
       }),
     },
+    propsSet: {
+      default: {
+        ContentGroup: {
+          heading: 'Natural language processing (NLP)',
+          showCopy: 'Copy:',
+          copy:
+            'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec quam ante, mattis id pellentesque at,' +
+            ' molestie et ipsum. Proin sodales est hendrerit maximus malesuada. Orci varius natoque penatibus et ' +
+            'magnis dis parturient montes, nascetur ridiculus mus. Etiam at arcu ligula. Praesent faucibus est ligula,' +
+            ' vitae finibus ante aliquet a.',
+          addChildren: '',
+          cta: true,
+        },
+      },
+    },
   },
 };