+++ packages/web-components/src/components/masthead/top-nav.ts
@@ -1,7 +1,7 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2020, 2021
+ * Copyright IBM Corp. 2020, 2022
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -183,9 +183,7 @@ class DDSTopNav extends StableSelectorMixin(HostListenerMixin(BXHeaderNav)) {
   private _handleSlotChange(event) {
     this._childItems = (event.target as HTMLSlotElement)
       .assignedNodes()
-      .filter(elem =>
-        (elem as HTMLElement).matches !== undefined ? (elem as HTMLElement).matches('dds-megamenu-top-nav-menu') : false
-      );
+      .filter(elem => (elem as HTMLElement).matches?.('dds-megamenu-top-nav-menu'));
     this._paginateRight({ onLoad: true });
   }